do not analyze tables or recompute db caches in db-migrate command
